Making changes to main,student and session class to make it more readable

+++ Exercise/exercise1/src/Student.java
@@ -3,7 +3,7 @@ import java.util.ArrayList;
 /*creating student class and we will be using this class to create an arraylist in session class*/
 class Student {
     private String student_name;
-    private ArrayList<Double> quiz_scores;
+    private ArrayList<Integer> quiz_scores;
 
     public Student(String student_name) {
         this.student_name = student_name;
@@ -12,18 +12,18 @@ class Student {
     }
 
     //calculates average quiz_scores
-    public double calculate_average_score_for_each_student() {
-        double sum = 0.0;
+    public float calculate_average_score_for_each_student() {
+        float sum = 0.0f;
         for (int i = 0; i < quiz_scores.size(); i++) {
             sum += quiz_scores.get(i);
         }
-        double average = sum / quiz_scores.size();
+        float average = sum / quiz_scores.size();
         return average;
 
     }
 
     // adds score to the quiz_scores arraylist when this method is called
-    public void score_add(double score) {
+    public void score_add(int score) {
         quiz_scores.add(score);
     }
 
@@ -32,7 +32,7 @@ class Student {
         return student_name;
     }
 
-    public ArrayList<Double> getQuiz_scores() {
+    public ArrayList<Integer> getQuiz_scores() {
         return quiz_scores;
     }
 }
